requirements.txt
appdirs==1.4.4
argcomplete==1.10.3
asgiref==3.4.1
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
compressed-rtf==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect==1.1.0
defusedxml==0.7.1
dill==0.3.4
Django==4.0.1
django-cors-headers==3.10.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
extract-msg==0.28.7
fake-useragent==0.1.11
idna==3.3
IMAPClient==2.1.0
importlib-metadata==4.10.0
itypes==1.2.0
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
multiprocess==0.70.12.2
numpy==1.22.0
oauthlib==3.1.1
olefile==0.46
pandas==1.3.5
parse==1.19.0
pdfminer.six==20191110
Pillow==9.0.0
pycparser==2.21
pycryptodome==3.12.0
pyee==8.2.2
PyJWT==2.3.0
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tabula-py==2.3.0
textract==1.6.4
tqdm==4.62.3
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
w3lib==1.22.0
websockets==9.1
xlrd==1.2.0
XlsxWriter==3.0.2
zipp==3.7.0
